1570 - Reorganize, SF
