+++ server/libs/platform/platform-ai/platform-ai-api/src/main/java/com/bytechef/platform/ai/AiCopilot.java
@@ -14,20 +14,16 @@
  * limitations under the License.
  */
 
-package com.bytechef.platform.ai.facade.dto;
+package com.bytechef.platform.ai;
 
-import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
+import com.bytechef.platform.ai.dto.ContextDTO;
 import java.util.Map;
+import reactor.core.publisher.Flux;
 
 /**
  * @author Ivica Cardic
  */
-@SuppressFBWarnings("EI")
-public record ContextDTO(Source source, Map<String, ?> parameters, String workflowId) {
+public interface AiCopilot {
 
-    public enum Source {
-        WORKFLOW_EDITOR,
-        WORKFLOW_EDITOR_COMPONENTS_POPOVER_MENU,
-        CODE_EDITOR
-    }
+    Flux<Map<String, ?>> chat(String message, ContextDTO contextDTO, String conversationId);
 }
