+++ server/libs/platform/platform-ai/platform-ai-api/src/main/java/com/bytechef/platform/ai/package-info.java
@@ -16,7 +16,7 @@
 
 @NonNullApi
 @NonNullFields
-package com.bytechef.platform.ai.facade;
+package com.bytechef.platform.ai;
 
 import org.springframework.lang.NonNullApi;
 import org.springframework.lang.NonNullFields;
